+++ saas_apps_signup/doc/changelog.rst
@@ -5,6 +5,7 @@
 - **Improvement:** in signup form language and country fields became mandatory
 - **Improvement:** moved SaaS settings to "SaaS" section in common settings
 - **Improvement:** added ability to set default number of days until expiration
+- **Fix:** correct template searching when user pressed 'Buy now'
 
 
 `0.1.0`